celebrityindian/requirements.txt
scrapy
mysql-connector-python
bs4